app/requirements.txt
argparse
python-docx
python-dotenv
beautifulsoup4
cohere
openai
pydantic
